refs #1149: fixed compiler warnings on VS x64 compilers
git-svn-id: https://svn.modelica.org/projects/Modelica/trunk@6570 7ce873d0-865f-4ce7-a662-4bb36ea78beb

+++ Modelica/Resources/C-Sources/win32_dirent.c
@@ -124,7 +124,7 @@ opendir (const char *szPath)
   /* Start the search, in order that the user may still
      change the current directory afterwards
   */
-  nd->dd_handle = _findfirst (nd->dd_name, &(nd->dd_dta));
+  nd->dd_handle = (long)_findfirst (nd->dd_name, &(nd->dd_dta));
   if (nd->dd_handle == -1) {
      /* Whoops! Seems there are no files in that
       * directory. */
@@ -192,7 +192,7 @@ readdir (DIR * dirp)
       /* Successfully got an entry. Everything about the file is
        * already appropriately filled in except the length of the
        * file name. */
-      dirp->dd_dir.d_namlen = strlen (dirp->dd_dir.d_name);
+      dirp->dd_dir.d_namlen = (unsigned short)strlen (dirp->dd_dir.d_name);
       return &dirp->dd_dir;
     }
